Removed yaml dependency -> jsonc
PyYaml is native in the arch installation of Python but not others.
As a result, we decided to remove this additional dependency and
simply use jsonc. The code and tests were updated to handle this
change.

Further changes were made to the README to be more verbose for manual
installation of the frontend.